+++ src/terrafort/templates/aws_security_group.import.j2
@@ -1,0 +1,1 @@
+terraform import aws_security_group.{{ resource.GroupName }} {{ resource.GroupId }}
